+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HIT lexer parameter key pattern so space is not required before override syntax
 - Diagnostic object message access to clean up string before returning downstream
 - CustomPlotFile so valid JSON is made from default rangeMin and rangeMax of axes
+- SnippetManager tabstop column offset to no longer include snippet syntax
 
 ### Changed
 - Minimum version of CMake to 3.20, removing deprecated version incompatibility/error observed with newer CMake